+++ lib/src/form/validators.dart
@@ -1,6 +1,12 @@
 import '../extensions/extensions.dart';
 
+/// {@template validator}
+///
+/// A [Validator] is base class for handling form validation.
+///
+/// {@endtemplate}
 abstract class Validator<T> {
+  /// {@macro validator}
   Validator(this.message);
   final String? message;
 
@@ -11,6 +17,11 @@ abstract class Validator<T> {
   }
 }
 
+/// {@template emailValidator}
+///
+/// A [Validator] that validates an email address.
+///
+/// {@endtemplate}
 class EmailValidator extends Validator<String> {
   EmailValidator(super.message);
 
@@ -23,13 +34,33 @@ class EmailValidator extends Validator<String> {
   bool isValid(String value) => value.isEmail;
 }
 
-class RequiredValidator extends Validator<String> {
+/// {@template requiredValidator}
+///
+/// A [Validator] that validates if a value is not null or empty.
+///
+/// {@endtemplate}
+class RequiredValidator<T> extends Validator<T> {
   RequiredValidator(super.message);
 
   @override
-  bool isValid(String value) => value.isNotBlank;
+  bool isValid(T value) {
+    if (value is String) {
+      return value.isNotNullNotEmpty;
+    } else if (value is List) {
+      return value.isNotNullNotEmpty;
+    } else if (value is Map) {
+      return value.isNotEmpty;
+    } else {
+      return value != null;
+    }
+  }
 }
 
+/// {@template maxLengthValidator}
+///
+/// A [Validator] that validates max length of string.
+///
+/// {@endtemplate}
 class MaxLengthValidator extends Validator<String> {
   final int limit;
 
@@ -39,6 +70,11 @@ class MaxLengthValidator extends Validator<String> {
   bool isValid(String value) => value.length <= limit;
 }
 
+/// {@template minLengthValidator}
+///
+/// A [Validator] that validates min length of string.
+///
+/// {@endtemplate}
 class MinLengthValidator extends Validator<String> {
   final int limit;
 
@@ -49,6 +85,14 @@ class MinLengthValidator extends Validator<String> {
     return value.length >= limit;
   }
 }
+
+/// {@template matchValidator}
+///
+/// A [Validator] that compares two values.
+/// If [compareFn] is not provided, it will compare the string values
+/// of the two values.
+///
+/// {@endtemplate}
 
 class MatchValidator<T> extends Validator<T> {
   MatchValidator(
@@ -65,14 +109,5 @@ class MatchValidator<T> extends Validator<T> {
   bool isValid(T value) {
     return compareFn?.call(value, match) ??
         value.toString().trim() == match.toString().trim();
-  }
-}
-
-class NullStringValidator extends Validator<String?> {
-  NullStringValidator(super.message);
-
-  @override
-  bool isValid(String? value) {
-    return value.isNotNullNotEmpty;
   }
 }