+++ lib/src/widgets/widgets.dart
@@ -4,4 +4,6 @@ export 'conditional_widget.dart';
 export 'hide_keyboard.dart';
 export 'row_with_space.dart';
 export 'scaffold_wrapper.dart';
+export 'scrollable_column.dart';
+export 'scrollable_row.dart';
 export 'social_button.dart';